requirements.txt
PyGithub
colorama
requests
websocket